+++ src/components/WidgetEditor.js
@@ -237,6 +237,18 @@ class WidgetEditor extends React.Component {
     if (this.props.widgetCaption !== nextProps.widgetCaption) {
       this.props.setCaption(nextProps.widgetCaption);
     }
+
+    // Whenever the title changes, we call the callback
+    if (this.props.widgetEditor.title !== nextProps.widgetEditor.title
+      && this.props.onChangeWidgetTitle) {
+      this.props.onChangeWidgetTitle(nextProps.widgetEditor.title);
+    }
+
+    // Whenever the caption changes, we call the callback
+    if (this.props.widgetEditor.caption !== nextProps.widgetEditor.caption
+      && this.props.onChangeWidgetCaption) {
+      this.props.onChangeWidgetCaption(nextProps.widgetEditor.caption);
+    }
   }
 
   componentDidUpdate(previousProps, previousState) {
@@ -938,18 +950,8 @@ class WidgetEditor extends React.Component {
         if (limit) this.props.setLimit(limit);
         if (chartType) this.props.setChartType(chartType);
         if (areaIntersection) this.props.setAreaIntersection(areaIntersection);
-        if (name) {
-          this.props.setTitle(name);
-          if (this.props.onChangeWidgetTitle) {
-            this.props.onChangeWidgetTitle(name);
-          }
-        }
-        if (caption) {
-          this.props.setCaption(caption);
-          if (this.props.onChangeWidgetCaption) {
-            this.props.onChangeWidgetCaption(caption);
-          }
-        }
+        if (name) this.props.setTitle(name);
+        if (caption) this.props.setCaption(caption);
         if (zoom) this.props.setZoom(zoom);
         if (lat && lng) this.props.setLatLng({ lat, lng });
         if (bbox) this.props.setBounds([[bbox[1], bbox[0]], [bbox[3], bbox[2]]]);
@@ -1016,9 +1018,6 @@ class WidgetEditor extends React.Component {
   handleTitleChange(event) {
     const title = event.target.value;
     this.props.setTitle(title);
-    if (this.props.onChangeWidgetTitle) {
-      this.props.onChangeWidgetTitle(title);
-    }
   }
 
   /**